requirements.txt
selenium
pytest==6.2.5
allure-pytest
